adding tasks to fetch marketing content. moving util methods to vector_search app
